ai-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
openai==1.12.0
python-dotenv==1.0.1
requests==2.31.0
pydantic==2.6.0